FEATURE: Added a config setting for verbose mode.
The config file get another section called "Other", the script will
not execute without this section present.
The only option in the section at this moment is the verbose boolean.

When the verbose boolean is set to True the script will start printing
helpful, but not overly verbose, information about the execution to
STDOUT.